Move optimizers and lr_schedulers to utils
